% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\name{AIC.regional_mix}
\alias{AIC.regional_mix}
\title{AIC from a regional_mix model}
\usage{
\method{AIC}{regional_mix}(object, ..., k = 2)
}
\arguments{
\item{object}{A regional_mix object}

\item{k}{AIC penality}

\item{\\dots}{extra things for AIC}
}
\description{
AIC from a regional_mix model
}
