% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_s3-class.R
\name{AIC.species_mix}
\alias{AIC.species_mix}
\title{Return AIC from a species_mix model}
\usage{
\method{AIC}{species_mix}(object, k = NULL, ...)
}
\arguments{
\item{object}{A species mix object}

\item{k}{AIC parameter}

\item{\\dots}{Ignored}
}
\description{
Return AIC from a species_mix model
}
