% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\name{BIC.regional_mix}
\alias{BIC.regional_mix}
\title{BIC from a regional_mix model}
\usage{
\method{BIC}{regional_mix}(object, ...)
}
\arguments{
\item{object}{A regional_mix object}

\item{\\dots}{Ignored}
}
\description{
BIC from a regional_mix model
}
