% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{BIC.species_mix}
\alias{BIC.species_mix}
\title{Return BIC from a species_mix model}
\usage{
\method{BIC}{species_mix}(object, ...)
}
\arguments{
\item{object}{A species mix object}

\item{\\dots}{Ignored}
}
\description{
Return BIC from a species_mix model
}
