% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap function for species archetype models}
\usage{
bootstrap(
  object,
  nboot = 10,
  type = "BayesBoot",
  mc.cores = 1,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{object}{A species_mix model object}

\item{nboot}{The number of bootstraps to run.}

\item{type}{the type of bootstrap to use, options are "SimpleBoot" which is a parametric bootstrap,
or "BayesBoot" which uses a Bayesian Bootstrap. "BayesBoot" is the default}

\item{mc.cores}{The number of cores to use when running a bootstrap. The default is 1.}

\item{quiet}{If TRUE, do not print progress of bootstrap.}
}
\description{
bootstrap function for species archetype models
}
