% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\name{coef.regional_mix}
\alias{coef.regional_mix}
\title{Print the coefficientf from a regional_mix model}
\usage{
\method{coef}{regional_mix}(object, ...)
}
\arguments{
\item{object}{A regional_mix object}

\item{\\dots}{Ignored}
}
\description{
Print the coefficientf from a regional_mix model
}
