% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_s3-class.R
\name{coef.species_mix}
\alias{coef.species_mix}
\title{print the coefficients from the species_mix model}
\usage{
\method{coef}{species_mix}(object, ...)
}
\arguments{
\item{object}{A species mix object}

\item{\\dots}{Ignored}
}
\description{
print the coefficients from the species_mix model
}
