% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics-class.R
\name{AIC}
\alias{AIC}
\alias{AIC.ecomix}
\title{diagnostics objects}
\usage{
\method{AIC}{ecomix}(object, ..., k = 2)
}
\arguments{
\item{object}{an ecomix object returned from \code{ecomix}.}
}
\description{
diagnostic tools for checking \code{ecomix} models; see \code{ecomix} for more specifics about how
run an ecomix model.
}
