% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\docType{package}
\name{ecomix-package}
\alias{ecomix-package}
\title{ecomix package}
\description{
ecomix is a package for implementing species archetype models (SAM) and region of
common profile models (RCP) in R. The two main functions are \code{species_mix} and \code{regional_mix}, 
which fits either a Species Archetype Model (SAM) or Region of Common Profile Model (RCP) based on input parameters.
}
