% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix_effects.R
\name{effect_data}
\alias{effect_data}
\alias{effect_data.regional_mix}
\title{Generate data for plotting or predicting partial effects of covariates}
\usage{
effect_data.regional_mix(focal.predictors, mod, ngrid = 50, ...)
}
\arguments{
\item{focal.predictors}{A character or string of characters which represent covariates in the model.}

\item{mod}{The fitted species_mix or regional_mix model.}

\item{ngrid}{The length of the prediction vector.}

\item{...}{other arguments}
}
\value{
This function should return a list of data.frames one for each focal.predictor.
This will enable user to predict marginal effects or plot the partial response plots.
}
\description{
This function produces a list of data.frames for predicting the partial
effect of a focal.predictor current included in a species_mix model.
}
\details{
Generate data for plotting or predicting partial effects of covariates

This function produces a list of data.frames for predicting the partial
effect of a focal.predictor current included species_mix or regional_mix models.
}
