% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix_s3-class.R
\name{extractAIC.regional_mix}
\alias{extractAIC.regional_mix}
\title{Extract AIC from regional_mix model.}
\usage{
\method{extractAIC}{regional_mix}(fit, scale = 1, k = 2, ...)
}
\arguments{
\item{fit}{Fitted RCP model}

\item{scale}{scale parameter}

\item{k}{AIC parameter}

\item{\\dots}{Ignored}
}
\description{
Extract AIC from regional_mix model.
}
