% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\name{ggplot.zonation}
\alias{ggplot.zonation}
\title{Plot a spatial RCP prediction}
\usage{
\method{ggplot}{zonation}(data, ...)
}
\arguments{
\item{data}{a zonation object.}

\item{...}{other arguments.}
}
\description{
Plot a spatial RCP model prediction using ggplot
}
\examples{
library(raster)
r1 <- raster(matrix(runif(200^2, 0, 1), 200))
r2 <- raster(matrix(runif(200^2, 0, 1), 200))
plan <- zonation(stack(r1, r2))
library(ggplot2)
ggplot(plan)

}
