% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix_s3-class.R
\name{logLik.regional_mix}
\alias{logLik.regional_mix}
\title{Extract log-likelihood from a regional_mix model.}
\usage{
\method{logLik}{regional_mix}(object, ...)
}
\arguments{
\item{object}{A fitted regional_mix model}

\item{\\dots}{Ignored}
}
\description{
Extract log-likelihood from a regional_mix model.
}
