% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_s3-class.R
\name{logLik.species_mix}
\alias{logLik.species_mix}
\title{Extract log-likelihood from a species_mix model.}
\usage{
\method{logLik}{species_mix}(object, ...)
}
\arguments{
\item{object}{A fitted species_mix model}

\item{\\dots}{Ignored}
}
\description{
Extract log-likelihood from a species_mix model.
}
