% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\name{plot.regional_mix.sampling}
\alias{plot.regional_mix.sampling}
\title{Plot residuals from a regional_mix model.}
\usage{
\method{plot}{regional_mix.sampling}(
  x,
  ...,
  object2 = NULL,
  sampling_levels,
  CI = c(0.025, 0.975),
  col = "black",
  lty = 1
)
}
\arguments{
\item{object2}{a bootstrap object from regional_mix.bootstrap}

\item{object}{a fitted regional_mix model you wish to plot}

\item{type}{What type of residuals to plot? 'RQR'; random quantile residuals or 'deviance' residuals.}

\item{nsim}{Number of simulations to run}

\item{alpha.conf}{The bounds of the confidence intervals.}

\item{quiet}{Run in quiet mode.}

\item{species}{Which species to plot as residuals.}

\item{fitted.scale}{What scale to plot the residuals on?}
}
\description{
Plot residuals from a regional_mix model.
}
\details{
The two types of residuals are inherently different. The "RQR" residuals produce a residual for each species at each site and the "deviance" residuals produce a site residual (no species level residual). The plots also differ, the "RQR" type generates a single normal QQ-plot for all species and all sites, and a residual versus fitted plot for all species and sites (Described in Foster et al, 2013). The "deviance" type generates a pair of Tukey mean-difference plots, similar in spirit to a QQ-plot. The first is for point-wise confidence intervals and the second is for approximate global intervals. See Foster et al (2013) for details.
The family for the "RQR" residuals should be standard normal. For "deviance" residuals, the distribution is unknown and simulation is used to graphically assess how odd the observed residuals look compared to ones generated assuming the model is correct.
}
\references{
Dunn, P.K. and Smyth G.K. (1996) Randomized Quantile Residuals. Journal of Computational and Graphical Statistics \emph{5}: 236--244.
Foster, S.D., Givens, G.H., Dornan, G.J., Dunstan, P.K. and Darnell, R. (2013) Modelling Regions of Common Profiles Using Biological and Environmental Data. Environmetrics \emph{24}: 489--499. DOI: 10.1002/env.2245
Foster, S.D., Hill, N.A. and Lyons, M., 2017. Ecological grouping of survey sites when sampling artefacts are present. Journal of the Royal Statistical Society: Series C (Applied Statistics), 66(5), pp.1031-1047.
}
