% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix_effects.R
\name{plot.regional_mix_effects_data}
\alias{plot.regional_mix_effects_data}
\title{plot.regional_mix_effects_data}
\usage{
\method{plot}{regional_mix_effects_data}(
  x,
  object,
  boot.object = NULL,
  nboot = 0,
  response.var = NULL,
  CI = c(0.025, 0.975),
  linecols = c("#1B9E77", "#D95F02", "#7570B3", "#E7298A", "#66A61E", "#E6AB02",
    "#A6761D", "#666666"),
  polycols = "#00000020",
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{a list of partial prediction data frames as generated by
effects_data.}

\item{object}{A fitted regional_mix model.}

\item{boot.object}{A regional_mix.bootstrap object. Default is NULL, no standard
errors will be reported.}

\item{nboot}{An option to do bootstrapping when plotting, this will be slow,
better to run and save a bootstrap object and pass to the plotting function as
boot.object.}

\item{response.var}{Which RCPs to plot on the y-axis. Default is all RCPs.
Can be a subset, using the names stored in \code{object$names$RCPs} (e.g. "RCP_1").}

\item{CI}{is the confidence intervals for the standard errors.}

\item{linecols}{Are the default colours for plotting the partial responses.}

\item{polycols}{Is the colour of the confidence intervals in the response plots.}

\item{ylim}{Default is NULL and will plot ylim within range of the response variable.}

\item{\\dots}{Extra plotting arguments.}
}
\description{
plot.regional_mix_effects_data
}
\details{
Plots the partial dependence plots (marginal response curves) for
focal covariates. Continuous covariates will be plotted as lines, factors
will be plotted as dotplots. RCP predictions are probabilities, so unlike
\link[ecomix]{plot.species_mix_effects_data} there is no link/response scale option.
}
