% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\name{plot.regional_mix_stab}
\alias{plot.regional_mix_stab}
\title{Diagnostic plotting to see if RCP groups are stable}
\usage{
\method{plot}{regional_mix_stab}(x, y, minWidth = 1, ncuts = 111, ylimmo = NULL, ...)
}
\arguments{
\item{x}{x-axis}

\item{y}{y-axis}

\item{minWidth}{min width of cuts/binning}

\item{ncuts}{number of cuts/bins to make}

\item{ylimmo}{limit of y-axis}

\item{\\dots}{additional plotting calls}
}
\description{
For increasing size of hold-out samples, cooks distance and predictive log-likelihood are plotted.
}
\examples{
\dontrun{
 #not run as R CMD check complains about the time taken.
 #This code will take a little while to run (about 3.5minutes on my computer)
 system.time(\{
 my.registab <- stability.regional_mix( fm, oosSizeRange=seq( from=1,to=fm$n\\%/\\%5,length=5),
                                     times=fm$n, mc.cores=2, doPlot=FALSE);
 plot( my.registab, minWidth=1, ncuts=15);
 \})
}
}
