% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_s3-class.R
\name{plot.species_membership}
\alias{plot.species_membership}
\title{plot.species_membership}
\usage{
\method{plot}{species_membership}(x, ...)
}
\arguments{
\item{x}{a fitted species_mix model.}

\item{\\dots}{Extra plotting arguments.}
}
\description{
plot.species_membership
}
\details{
Plot the posterior taus as a heatmap. Look at
}
