% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_s3-class.R
\name{plot.species_mix.multifit}
\alias{plot.species_mix.multifit}
\title{plot.species_mix.multifit}
\usage{
\method{plot}{species_mix.multifit}(x, type = c("BIC", "AIC", "logLik"), ...)
}
\arguments{
\item{x}{a fitted species_mix.multifit object.}

\item{type}{What type of response to plot options are "BIC", "AIC" or "logLik".}

\item{\\dots}{Extra plotting arguments.}
}
\description{
plot.species_mix.multifit
}
\details{
Plot BIC or other likelihood based indices from multiple fit object.
}
