% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{plot.species_mix}
\alias{plot.species_mix}
\title{plot.species_mix}
\usage{
\method{plot}{species_mix}(x, species = "AllSpecies", fitted.scale = "response", ...)
}
\arguments{
\item{x}{a fitted species_mix model.}

\item{species}{which species residuals to plot. Default is "AllSpecies".}

\item{fitted.scale}{log or logit, this enables the plotting of residuals to be on the linear predictor scale.}

\item{\\dots}{Extra plotting arguments.}
}
\description{
plot.species_mix
}
\details{
Plot random quantile residuals (RQR). "RQR" produces residuals for each species.
}
\references{
Dunn, P.K. and Smyth G.K. (1996) Randomized Quantile Residuals. Journal of Computational and Graphical Statistics \emph{5}: 236--244.
}
