% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix_s3-class.R
\name{print.regional_mix.multifit}
\alias{print.regional_mix.multifit}
\title{Prints a summary of a regional_mix.multifit object.}
\usage{
\method{print}{regional_mix.multifit}(x, ...)
}
\arguments{
\item{x}{A regional_mix.multifit object}

\item{\\dots}{Ignored}
}
\description{
Reports the BIC of each of the random starts and identifies the best fitting model.
}
