% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix_s3-class.R
\name{print.regional_mix}
\alias{print.regional_mix}
\title{Prints some attributes of a regimix object.}
\usage{
\method{print}{regional_mix}(x, ...)
}
\arguments{
\item{x}{A fitted regional_mix object}

\item{\\dots}{Ignored}
}
\description{
A list is returned that will be printed on exit, if not assigned to anything. It contains the function call and the estimated coefficients.
}
