% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{print.species_mix}
\alias{print.species_mix}
\title{Print a species_mix model object.}
\usage{
\method{print}{species_mix}(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{\\dots}{Ignored}
}
\description{
Print a species_mix model object.
}
\examples{

#Print information about a species_mix model
\dontrun{
print(fm1)}
}
