% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix-class.R
\name{regional_mix.species_membership}
\alias{regional_mix.species_membership}
\title{What is the average species membership per RCP?}
\usage{
regional_mix.species_membership(
  object,
  object2 = NULL,
  CI = c(0.025, 0.975),
  ...
)
}
\arguments{
\item{object}{A RCP model}

\item{object2}{A RCP model bootstrap object}

\item{CI}{The confidence intervals to report the range of values form bootstrap}

\item{\\dots}{Ignored for now.}
}
\description{
Extracts the average species' membership for each RCP.
}
