% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{residuals.species_mix}
\alias{residuals.species_mix}
\title{Estimate residuals for a species_mix object}
\usage{
\method{residuals}{species_mix}(object, ..., type = "RQR", control = species_mix.control())
}
\arguments{
\item{object}{A returned species_mix model object.}

\item{\dots}{additional calls for residual function}

\item{type}{The type of residuals to estimate. Default is "RQR" (Random Quantile Residuals).
But you can also simulate many Random Quantile Residuals using "SimRQR".}

\item{control}{Default uses species_mix.control()}
}
\description{
The randomised quantile residuals ("RQR", from Dunn and Smyth, 1996) are defined by
their marginal distribution function (marginality is over other species observations within that site;
see Woolley et al, in prep).
}
