% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_utils.R
\name{rGMRF}
\alias{rGMRF}
\title{Simulate a gaussian markov random field}
\usage{
rGMRF(x, y = x, sig2 = 1, rho = 0.5, nu = 1/2, nugget = NULL)
}
\arguments{
\item{x}{vector of evenly spaced values in the x-direction}

\item{y}{vector of evenly spaced values in the y-direction (default is the same as x)}

\item{sig2}{gives the marginal variance of the process}

\item{rho}{is effective range/lengthscale}

\item{nu}{is smoothness for the Matern, can be either 1/2, 3/2, 5/2 if none of these then an error will be thrown.}

\item{nugget}{is nugget variance. Default is NULL.}
}
\description{
simulate a Gaussian Markov Random Field from the Matern family.
This is developed from the Rue & Held 2005 Algorithim 2.10.
}
\examples{
# exponential with no nugget
x <- y <- seq(-2, 2, len=250)
covsim1 <- rGMRF(x,y, sig2 = 10,rho = .5, nu = 1/2)
image(x,y,covsim1)
# 3/2 Matern with nugget
x <- y <- seq(-2, 2, len=250)
covsim2 <- rGMRF(x,y,sig2 = .1,rho = .5, nu = 3/2, nugget = 1.5^2)
image(x,y,covsim2)
}
