% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_s3-class.R
\name{species_membership}
\alias{species_membership}
\title{Extract the posterior species membership per archetype.}
\usage{
species_membership(object, ...)
}
\arguments{
\item{object}{A species_mix model object}

\item{\\dots}{Ignored}
}
\description{
Extracts tau_{jk} from species_mix model object. These are the
posterior memberships of each species to each archetypes. Rows should sum
to 1.
}
