% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{species_mix.bootstrap}
\alias{species_mix.bootstrap}
\title{species_mix.bootstrap}
\usage{
species_mix.bootstrap(
  object,
  nboot = 1000,
  type = "BayesBoot",
  mc.cores = 1,
  quiet = FALSE
)
}
\arguments{
\item{object}{A species_mix model object}

\item{nboot}{The number of boostraps to run.}

\item{type}{the type of bootstrap to use, options are  "SimpleBoot" which is  a parameteric bootstrap, or "BayesBoot"}

\item{mc.cores}{The number of core to fit. The default is 1.}

\item{quiet}{If TRUE, do not print progress of bootstrap.}
}
\description{
species_mix.bootstrap
}
