% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_mix_s3-class.R
\name{summary.regional_mix}
\alias{summary.regional_mix}
\title{A summary from a regional_mix object.}
\usage{
\method{summary}{regional_mix}(object, ...)
}
\arguments{
\item{object}{A regional_mix model object}

\item{\\dots}{ignored.}
}
\description{
A summary from a regional_mix object.
}
\details{
A table is printed that contains the coefficient values, their standard errors, and their z-statistic. The second and thrid columns may be unreliable for some parameters.
}
