% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{summary.species_mix}
\alias{summary.species_mix}
\title{Print a summary of the species_mix model}
\usage{
\method{summary}{species_mix}(object, ...)
}
\arguments{
\item{object}{A species_mix model object}

\item{\\dots}{Ignored}
}
\description{
Print a summary of the species_mix model, this requires the variance-covariance matrix to be appended to the species_mix model.
}
