% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_s3-class.R
\name{terms.species_mix}
\alias{terms.species_mix}
\title{Return terms from a fitted species_mix model}
\usage{
\method{terms}{species_mix}(x, ...)
}
\arguments{
\item{x}{A species_mix model object}

\item{\\dots}{Ignored}
}
\description{
This function returns a list of terms objects, one for each formula in the species mix model.
\describe{
 \item{xterms}{Are the terms for the archetypes formula.}
 \item{wterms}{Are the terms for the species formula.}
 \item{uterms}{Are the terms for the all/bias formula.}
}
}
