% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix_bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrapping of species_mix or regional_mix fitted models.}
\usage{
bootstrap(
  object,
  nboot = 10,
  type = "BayesBoot",
  mc.cores = 1,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{object}{A species_mix model object}

\item{nboot}{The number of bootstraps to run.}

\item{type}{the type of bootstrap to use, options are "SimpleBoot" which is a parametric bootstrap,
or "BayesBoot" which uses a Bayesian Bootstrap. "BayesBoot" is the default}

\item{mc.cores}{The number of cores to use when running a bootstrap. The default is 1.}

\item{quiet}{If TRUE, do not print progress of bootstrap.}

\item{\\dots}{Ignored}
}
\value{
An object of class "species_mix.bootstrap", which is essentially a matrix with nboot rows and the number of columns equal to the number of parameters matrix. Each row gives a bootstrap estimate of the parameters.
}
\description{
This function can take a while to run -- it is a bootstrap function. nboot re-samples of the data are taken and then the parameters are estimated for each re-sample. The function allows for parallel calculations, via mclapply(qv), which reduces some of the computational burden. To use parallel computing, specify mc.cores>1. Note that this will not work on Windows computers, as mclapply(qv) will not work.
The Bayesian bootstrap method is operationally equivalent to the simple bootstrap, except that the weightings are non-integral. See Rubin (1981). It might be tempting to reduce the tolerance for convergence of each estimation procedure. We recommend not doing this as it is likely to have the effect of artificially reducing estimates of uncertainty. This occurs as the resampled estimates are likely to be closer to their starting values (the MLEs from the original data set).
}
\references{
Rubin, D.B. (1981) The Bayesian Bootstrap. The Annals of Statistics \emph{9}:130--134.
}
