% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ecomix}
\alias{ecomix}
\alias{ecomix-package}
\title{package file}
\description{
ecomix is a package for implementing species archetype models (SAM) and region of
  common profile models (RCP) in R. The main function is \code{ecomix}, which fits either a SAM or RCP
  based on input parameters.
}
