% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{predict.species_mix}
\alias{predict.species_mix}
\title{Predict a species_mix model.}
\usage{
\method{predict}{species_mix}(
  object,
  object2 = NULL,
  newdata = NULL,
  offset = NULL,
  nboot = 0,
  alpha = 0.95,
  mc.cores = 1,
  prediction_type = "archetype",
  ...
)
}
\arguments{
\item{object}{is a matrix model returned from the species_mix model.}

\item{object2}{is a species mix bootstrap object.}

\item{newdata}{a matrix of new observations for prediction.}

\item{offset}{an offset for prediction}

\item{nboot}{Number of bootstraps (or simulations if using IPPM) to run if no object2 is provided.}

\item{alpha}{confidence level. default is 0.95}

\item{mc.cores}{number of cores to use in prediction. default is 1.}

\item{prediction_type}{Do you want to produce 'archetype' or 'species' level predictions. default is 'archetype'.}

\item{\\dots}{Ignored}
}
\description{
Predict species archetypes from a species_mix model. You can also predict the conditional species predictions using "prediction_type='species'".
}
\examples{
\donttest{
library(ecomix)
set.seed(42)
sam_form <- stats::as.formula(paste0('cbind(',paste(paste0('spp',1:20),
collapse = ','),")~x1+x2"))
sp_form <- ~ 1
beta <- matrix(c(-2.9,-3.6,-0.9,1,.9,1.9),3,2,byrow=TRUE)
dat <- data.frame(y=rep(1,100),x1=stats::runif(100,0,2.5),
x2=stats::rnorm(100,0,2.5))
dat[,-1] <- scale(dat[,-1])
simulated_data <- species_mix.simulate(archetype_formula=sam_form,
 species_formula=sp_form, dat,beta=beta,dist="bernoulli")
fm1 <- species_mix(sam_form, sp_form, simulated_data,
 distribution = 'bernoulli',  n_mixtures=3)
preds_fm1 <- predict(fm1)
}
}
