% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_mix-class.R
\name{species_mix.simulate}
\alias{species_mix.simulate}
\title{Simulate species mix data for model fitting.}
\usage{
species_mix.simulate(
  archetype_formula,
  species_formula,
  all_formula = NULL,
  data,
  offset = NULL,
  nArchetypes = 3,
  alpha = NULL,
  beta = NULL,
  gamma = NULL,
  delta = NULL,
  logTheta = NULL,
  powers = NULL,
  size = NULL,
  family = "bernoulli"
)
}
\arguments{
\item{archetype_formula}{formula to simulate species_mix data, needs to have
the format: cbind(spp1,spp2,spp3,...,sppN)~1 + x1 + x2}

\item{species_formula}{formula to simulate species_mix species-specific
responses, e.g: ~1}

\item{offset}{used to offset sampling effort for abundance data (log link function).}

\item{nArchetypes}{number of groups to simulate.}

\item{alpha}{coefficients for each species archetype. vector S long.}

\item{beta}{coefficients for each species archetype. Matrix of G x number of
parameters. Each row is a different species archetype.}

\item{gamma}{coefficients for each species archetype. Matrix of S x number of
parameters. Each row is a different species archetype.}

\item{size}{Is for the binomial model and this represents the number of binomial trials per site, can be fixed or vary.}

\item{family}{Which statistical distribution to simulate data for.
'bernoulli','binomial', 'gaussian', 'ippm', 'negative.binomial' and 'poisson'.}

\item{dat}{a matrix of variables to simulate data from.}

\item{theta}{coefficients for the dispersion variables for negative.binomial
and gaussian distributions - should be number of species long}
}
\description{
Simulate species mix data for model fitting.
}
\examples{
\donttest{
archetype_formula <- stats::as.formula(paste0('cbind(',paste(paste0('spp',
1:20),collapse = ','),")~1+x1+x2"))
species_formula <- stats::as.formula(~1)
beta <- matrix(c(-3.6,0.5,
                 -0.9,1.0,
                  0.9,-2.9,
                  2.2,5.4),
                4,2,byrow=TRUE)
dat <- data.frame(y=rep(1,100),
                  x1=stats::runif(100,0,2.5),
                  x2=stats::rnorm(100,0,2.5))
simulated_data <- species_mix.simulate(archetype_formula,species_formula,
                                       dat, nArchetypes = 4, beta=beta,
                                       family="bernoulli")
}
}
